Apply browser-targeted WordPress patches via Blueprints, not in remote.html.
Enables patching arbitrary WordPress bundles for compatibility with web
browsers. Without this PR, remote.html applies these patch right after
the WordPress module is loaded into the /wordpress VFS directory.
However, in #700 we want to enable using an arbitrary WordPress build
from any PR in wordpress-develop. We must, thus, expose a way of
applying the necessary patches.

Testing instructions

1. Start Playground, log out
2. Go to /wp-login.php
3. Confirm there's a message explaining what login credentials to use
