--- packages/playground/remote/src/lib/web-wordpress-patches/wp-content/mu-plugins/includes/requests_transport_dummy.php
@@ -1,43 +1,0 @@
-<?php
-
-/**
- * This transport does not perform any HTTP requests and only exists
- * to prevent the Requests class from complaining about not having any
- * transports.
- */
-class Requests_Transport_Dummy implements Requests_Transport
-{
-	public $headers = '';
-
-	public function __construct()
-	{
-	}
-
-	public function __destruct()
-	{
-	}
-
-	public function request($url, $headers = array(), $data = array(), $options = array())
-	{
-		return false;
-	}
-
-	public function request_multiple($requests, $options)
-	{
-		$responses = array();
-		foreach ($requests as $id => $request) {
-			$responses[] = false;
-		}
-		return $responses;
-	}
-
-	protected static function format_get($url, $data)
-	{
-		return $url;
-	}
-
-	public static function test($capabilities = array())
-	{
-		return true;
-	}
-}
