+++ packages/playground/remote/src/lib/worker-thread.ts
@@ -13,7 +13,6 @@ import {
 	SupportedPHPVersion,
 	SupportedPHPVersionsList,
 } from '@php-wasm/universal';
-import { applyWebWordPressPatches } from './web-wordpress-patches';
 import {
 	SyncProgressCallback,
 	bindOpfs,
@@ -175,7 +174,6 @@ try {
 		 * already applied.
 		 */
 		await wordPressModule;
-		applyWebWordPressPatches(php);
 		await applyWordPressPatches(php, {
 			wordpressPath: DOCROOT,
 			patchSecrets: true,
